% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorlist-package.R
\docType{package}
\name{authorlist-package}
\alias{authorlist}
\alias{authorlist-package}
\title{authorlist: To Creat the Formatted Author List for Manuscripts with Many Authors}
\description{
Formatting the author list and the affiliation list in Word for the manuscript with many coauthors is painful and time-consuming. However, putting the author names and affiliations together in a CSV file is easy. The package "authorlist" automatically produces a Word document with the formatted author and affiliation list ready for submission from a CSV file. The template of the CSV file, which includes the coauthors' names, affiliations, and their order in the author list, can be downloaded from https://figshare.com/articles/dataset/CSV_template_for_R_package_authorlist_/16794901. The leading author of the manuscript could make it even easier by sharing the CSV template through Dropbox or Google Drive and asking the coauthors to fill in their information.
}
\author{
\strong{Maintainer}: Qiang Yang \email{qiang.yang@uni-konstanz.de} (\href{https://orcid.org/0000-0003-4210-9007}{ORCID})

}
\keyword{internal}
