% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{getTemplate}
\usage{
getTemplate(destfile = "authorlist.template.csv")
}
\arguments{
\item{destfile}{the path to save the CSV template, Default: 'authorlist.template.csv'}
}
\value{
the CSV template
}
\description{
download the CSV template
}
\details{
This function downloads the CSV template to your computer
}
\examples{
getTemplate(destfile = "~/Downloads/authorlist.template.csv")
}
