% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toAuthorlist.R
\name{toAuthorlist}
\alias{toAuthorlist}
\title{toAuthorlist}
\usage{
toAuthorlist(
  source.csv = "authorlist.csv",
  target = "authorlist.docx",
  type = "numbered"
)
}
\arguments{
\item{source.csv}{the path to the CSV document including the coauthor information, Default: 'authorlist.csv'}

\item{target}{the path where you would like to save the produced Word document, Default: 'authorlist.docx'}

\item{type}{the type of the affiliation numbering, could be "numbered" or "alphabeta", Default: 'numbered'. If 'numbered', the affiliation will be marked with numbers, otherwise the affiliation will be marked with letters as PNAS papers.}
}
\value{
the produced Word document
}
\description{
main function to produce the formatted author list and affiliation list in Word
}
\details{
This is the main function to produce the formatted author list and affiliation list in Word.
}
\examples{
toAuthorlist(source.csv = "~/Downloads/authorlist.csv",
           target = "~/Downloads/authorlist.docx",
           type="numbered")
}
